Initiated new expo project with typescript
