--- README.md
@@ -1,1 +1,0 @@
-# CoolGuides for Anything and Everything
