--- eas.json
@@ -1,18 +1,0 @@
-{
-  "cli": {
-    "version": ">= 0.35.0"
-  },
-  "build": {
-    "development": {
-      "developmentClient": true,
-      "distribution": "internal"
-    },
-    "preview": {
-      "distribution": "internal"
-    },
-    "production": {}
-  },
-  "submit": {
-    "production": {}
-  }
-}
