+++ package.json
@@ -1,60 +1,26 @@
 {
+  "name": "coolguides",
+  "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
-  "name": "cool-guides",
-  "version": "0.1.0",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "~1.15.0",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "@react-navigation/native": "^6.0.6",
-    "@react-navigation/native-stack": "^6.2.5",
-    "@react-navigation/web": "^1.0.0-alpha.9",
-    "dotenv": "^10.0.0",
-    "expo": "^43.0.0",
-    "expo-constants": "~12.1.3",
-    "expo-linking": "~2.4.2",
-    "expo-status-bar": "~1.1.0",
-    "firebase": "8.2.3",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-native": "0.64.2",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-image-zoom-viewer": "^3.0.1",
-    "react-native-paper": "^4.9.2",
-    "react-native-polyfill": "^1.2.1",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.8.0",
-    "react-native-web": "0.17.1",
-    "react-query": "^3.33.2"
+    "expo": "~46.0.9",
+    "expo-status-bar": "~1.4.0",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-native": "0.69.5",
+    "react-native-web": "~0.18.7"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
-    "@expo/webpack-config": "^0.16.2",
-    "@types/react": "~17.0.21",
-    "@types/react-native": "~0.64.12",
-    "babel-preset-expo": "8.5.1",
-    "typescript": "~4.3.5",
-    "workbox-background-sync": "^6.1.5",
-    "workbox-broadcast-update": "^6.1.5",
-    "workbox-cacheable-response": "^6.1.5",
-    "workbox-core": "^6.1.5",
-    "workbox-expiration": "^6.1.5",
-    "workbox-google-analytics": "^6.1.5",
-    "workbox-navigation-preload": "^6.1.5",
-    "workbox-precaching": "^6.1.5",
-    "workbox-range-requests": "^6.1.5",
-    "workbox-routing": "^6.1.5",
-    "workbox-strategies": "^6.1.5",
-    "workbox-streams": "^6.1.5",
-    "workbox-webpack-plugin": "^6.1.5"
+    "@types/react": "~18.0.14",
+    "@types/react-native": "~0.69.1",
+    "typescript": "~4.3.5"
   },
   "private": true
 }
