--- src/config/constants.jsx
@@ -1,11 +1,0 @@
-import { DefaultTheme } from "react-native-paper"
-
-export const theme = {
-  ...DefaultTheme,
-  roundness: 2,
-  colors: {
-    ...DefaultTheme.colors,
-    primary: "#1de9b6",
-    accent: "#E9F7E9",
-  },
-}
