--- src/navigation/AuthenticatedUserProvider.jsx
@@ -1,13 +1,0 @@
-import React, { useState, createContext } from "react"
-
-export const AuthenticatedUserContext = createContext({})
-
-export const AuthenticatedUserProvider = ({ children }) => {
-  const [user, setUser] = useState(null)
-
-  return (
-    <AuthenticatedUserContext.Provider value={{ user, setUser }}>
-      {children}
-    </AuthenticatedUserContext.Provider>
-  )
-}
